+++ CHANGES.rst
@@ -1,7 +1,12 @@
 
-Revision 0.1.5, XX-10-2017
+Revision 0.2.1, XX-11-2017
 --------------------------
 
+- Added MIB *status*, *product release* and *revision description* set
+  calls at pysnmp code generator
+- Changed REVISION field format in JSON representation - it is now
+  a list of dicts each with *revision* timestamp and *description* text
+- MIB REFERENCE fields are only exported if --with-mib-text is on
 - Sphinx documentation theme changed to Alabaster
 - Multiple fixes to pysnmp codegen not to produce function calls
   with more than 255 parameters